Renamed internal `notebooks` module to `workspace`
Fix #59
