+++ src/databricks/labs/pytester/fixtures/plugin.py
@@ -26,7 +26,7 @@ from databricks.labs.pytester.fixtures.catalog import (
     make_storage_credential,
     make_volume,
 )
-from databricks.labs.pytester.fixtures.notebooks import make_directory, make_workspace_file, make_notebook, make_repo
+from databricks.labs.pytester.fixtures.workspace import make_directory, make_workspace_file, make_notebook, make_repo
 from databricks.labs.pytester.fixtures.permissions import (  # noqa
     make_cluster_policy_permissions,
     make_instance_pool_permissions,
